Refactored waypoint/itinerary/time util functions
