+++ src/components/ITSBStore/index.js
@@ -1,1 +1,2 @@
+export * from './utils';
 export * from './ITSBStore';
